# Improve SSSE3 and AVX2 decoding speed