--- a/lib/codec_ssse3.c
+++ b/lib/codec_ssse3.c
@@ -91,23 +91,18 @@ enc_translate (const __m128i in)
 static inline __m128i
 dec_reshuffle (__m128i in)
 {
-	// Shuffle bytes to 32-bit bigendian:
-	in = _mm_bswap_epi32(in);
-
 	// Mask in a single byte per shift:
-	__m128i mask = _mm_set1_epi32(0x3F000000);
+	const __m128i maskB2 = _mm_set1_epi32(0x003F0000);
+	const __m128i maskB1 = _mm_set1_epi32(0x00003F00);
 
 	// Pack bytes together:
-	__m128i out = _mm_slli_epi32(_mm_and_si128(in, mask), 2);
-	mask = _mm_srli_epi32(mask, 8);
+	__m128i out = _mm_srli_epi32(in, 16);
 
-	out = _mm_or_si128(out, _mm_slli_epi32(_mm_and_si128(in, mask), 4));
-	mask = _mm_srli_epi32(mask, 8);
+	out = _mm_or_si128(out, _mm_srli_epi32(_mm_and_si128(in, maskB2), 2));
 
-	out = _mm_or_si128(out, _mm_slli_epi32(_mm_and_si128(in, mask), 6));
-	mask = _mm_srli_epi32(mask, 8);
+	out = _mm_or_si128(out, _mm_slli_epi32(_mm_and_si128(in, maskB1), 12));
 
-	out = _mm_or_si128(out, _mm_slli_epi32(_mm_and_si128(in, mask), 8));
+	out = _mm_or_si128(out, _mm_slli_epi32(in, 26));
 
 	// Reshuffle and repack into 12-byte output format:
 	return _mm_shuffle_epi8(out, _mm_setr_epi8(